Begin cutting old hookgroup pattern out
