+++ Content/Biomes/MoonstoneBiome.cs
@@ -1,14 +1,16 @@
-﻿//TODO ON WORLDGEN:
-//Wall updates
-//General cleanup of post-shape stuff
-
+﻿using StarlightRiver.Content.Items.Utility;
 using StarlightRiver.Content.Tiles.Moonstone;
 using StarlightRiver.Core.Loaders;
+using StarlightRiver.Core.Systems.CameraSystem;
 using StarlightRiver.Core.Systems.ScreenTargetSystem;
 using System;
+using System.Linq;
 using System.Reflection;
+using Terraria.DataStructures;
 using Terraria.Graphics.Effects;
 using Terraria.Graphics.Shaders;
+using Terraria.ID;
+using Terraria.ModLoader.IO;
 
 namespace StarlightRiver.Content.Biomes
 {
@@ -58,6 +60,9 @@ namespace StarlightRiver.Content.Biomes
 
 		private bool drawingBGtarget = false;
 
+		public bool moonstoneForced;
+		public bool meteorForced;
+
 		public ParticleSystem particleSystem;
 		public ParticleSystem particleSystemMedium;
 		public ParticleSystem particleSystemLarge;
@@ -88,6 +93,7 @@ namespace StarlightRiver.Content.Biomes
 
 			On_Main.DrawBackgroundBlackFill += DrawParticleTarget;
 			On_Main.DrawSurfaceBG += DistortBG;
+			On_WorldGen.meteor += ReplaceMeteorWithMoonstone;
 		}
 
 		public override void PostUpdateEverything()
@@ -244,6 +250,111 @@ namespace StarlightRiver.Content.Biomes
 			Main.spriteBatch.Begin(default, default, Main.DefaultSamplerState, default, Main.Rasterizer, default, Main.GameViewMatrix.TransformationMatrix);
 		}
 
+		/// <summary>
+		/// If the next meteor-like event to spawn should me a moonstone or not
+		/// </summary>
+		/// <returns></returns>
+		private bool ShouldBeMoonstone()
+		{
+			if (moonstoneForced)
+			{
+				moonstoneForced = false;
+				return true;
+			}
+
+			if (meteorForced)
+			{
+				meteorForced = false;
+				return false;
+			}
+
+			return Main.rand.NextBool();
+		}
+
+		private bool ReplaceMeteorWithMoonstone(On_WorldGen.orig_meteor orig, int i, int j, bool ignorePlayers)
+		{
+			CameraSystem.shake += 80;
+			Terraria.Audio.SoundEngine.PlaySound(SoundID.DD2_ExplosiveTrapExplode);
+
+			if (ShouldBeMoonstone())
+			{
+				var target = new Point16();
+
+				while (!SafeToSpawnMoonstone(target))
+				{
+					int x = Main.rand.Next(Main.maxTilesX);
+
+					for (int y = 0; y < Main.maxTilesY; y++)
+					{
+						if (Framing.GetTileSafely(x, y).HasTile)
+						{
+							target = new Point16(x, y - 20);
+							break;
+						}
+					}
+				}
+
+				for (int x = -10; x < 10; x++)
+				{
+					for (int y = -30; y < 30; y++)
+					{
+						if (Math.Abs(x) < 10 - Math.Abs(y) / 3 + StarlightWorld.genNoise.GetPerlin(x * 4, y * 4) * 8)
+							WorldGen.PlaceTile(target.X + x, target.Y + y, ModContent.TileType<Tiles.Moonstone.MoonstoneOre>(), true, true);
+					}
+				}
+
+				for (int x = -15; x < 15; x++)
+				{
+					for (int y = 0; y < 40; y++)
+					{
+						if (Math.Abs(x) < 10 - Math.Abs(y) / 3 + StarlightWorld.genNoise.GetPerlin(x * 4, y * 4) * 8)
+							WorldGen.PlaceTile(target.X + x, target.Y + y, ModContent.TileType<Tiles.Moonstone.MoonstoneOre>(), true, true);
+					}
+				}
+
+				Terraria.Chat.ChatHelper.BroadcastChatMessage(NetworkText.FromLiteral("A shard of the moon has landed!"), new Color(107, 233, 231));
+
+				if (Main.netMode == NetmodeID.Server)
+					NetMessage.SendTileSquare(Main.myPlayer, target.X - 30, target.Y - 30, 60, 70, TileChangeType.None);
+
+				return true;
+			}
+			else
+			{
+				return orig(i, j, ignorePlayers);
+			}
+		}
+
+		/// <summary>
+		/// Emulates teh various safety checks vanilla uses for if a meteor can be spawned
+		/// </summary>
+		/// <param name="test">The center point that a moonstone is attempting to be spawned at</param>
+		/// <returns></returns>
+		private bool SafeToSpawnMoonstone(Point16 test)
+		{
+			if (test == Point16.Zero)
+				return false;
+
+			for (int x = -35; x < 35; x++)
+			{
+				for (int y = -35; y < 35; y++)
+				{
+					if (WorldGen.InWorld(test.X + x, test.Y + y))
+					{
+						Tile tile = Framing.GetTileSafely(test + new Point16(x, y));
+
+						if (tile.TileType == TileID.Containers || tile.TileType == TileID.Containers2)
+							return false;
+					}
+				}
+			}
+
+			if (Main.npc.Any(n => n.active && n.friendly && Vector2.Distance(n.Center, test.ToVector2() * 16) <= 35 * 16))
+				return false;
+			else
+				return true;
+		}
+
 		protected void UpdateMoonParticles(Particle particle)
 		{
 			float parallax = 0.6f;
@@ -264,6 +375,18 @@ namespace StarlightRiver.Content.Biomes
 			Color color = Color.White;
 			particle.Color = color * opacity * fade;
 			particle.Timer--;
+		}
+
+		public override void SaveWorldData(TagCompound tag)
+		{
+			tag.Add("moonstoneForced", moonstoneForced);
+			tag.Add("meteorForced", meteorForced);
+		}
+
+		public override void LoadWorldData(TagCompound tag)
+		{
+			moonstoneForced = tag.GetBool("moonstoneForced");
+			meteorForced = tag.GetBool("meteorForced");
 		}
 	}
 }
