+++ Content/Bosses/SquidBoss/NPCs.ArenaBlocker.cs
@@ -12,19 +12,9 @@ namespace StarlightRiver.Content.Bosses.SquidBoss
 
 		public override string Texture => AssetDirectory.Invisible;
 
-		public override bool PreDraw(SpriteBatch spriteBatch, Vector2 screenPos, Color drawColor)
+		public override void Load()
 		{
-			return false;
-		}
-
-		public override bool CanHitPlayer(Player target, ref int cooldownSlot)
-		{
-			return State == 0;
-		}
-
-		public override bool CheckActive()
-		{
-			return !NPC.AnyNPCs(NPCType<SquidBoss>());
+			On_Player.Update_NPCCollision += ForceColliding;
 		}
 
 		public override void SetDefaults()
@@ -44,6 +34,26 @@ namespace StarlightRiver.Content.Bosses.SquidBoss
 			database.Entries.Remove(bestiaryEntry);
 		}
 
+		public override bool CanHitPlayer(Player target, ref int cooldownSlot)
+		{
+			return State == 0;
+		}
+
+		public override bool CheckActive()
+		{
+			return !NPC.AnyNPCs(NPCType<SquidBoss>());
+		}
+
+		public override void ModifyHitPlayer(Player target, ref Player.HurtModifiers modifiers)
+		{
+			target.immune = true;
+			target.immuneTime = 1;
+
+			target.position.Y = NPC.position.Y + NPC.height;
+			target.velocity.Y += 5;
+			target.noKnockback = true;
+		}
+
 		public override void AI()
 		{
 			if (!NPC.AnyNPCs(NPCType<SquidBoss>()))
@@ -60,6 +70,39 @@ namespace StarlightRiver.Content.Bosses.SquidBoss
 				if (proj.active && proj.Hitbox.Intersects(NPC.Hitbox) && proj.aiStyle == ProjAIStyleID.Hook) //Hooks
 					Main.projectile[k].active = false;
 			}
+		}
+
+		private void ForceColliding(On_Player.orig_Update_NPCCollision orig, Player self)
+		{
+			foreach(NPC npc in Main.ActiveNPCs)
+			{
+				int specialHit = 0;
+				float damageMult = 0f;
+				Rectangle npcRect = default;
+
+				if (npc.type == NPCType<ArenaBlocker>() && NPCLoader.CanHitPlayer(NPC, self, ref specialHit))
+				{
+					NPC.GetMeleeCollisionData(self.Hitbox, npc.whoAmI, ref specialHit, ref damageMult, ref npcRect);
+
+					if (self.Hitbox.Intersects(npcRect))
+					{
+						var hit = new Player.HurtInfo() { Damage = NPC.damage };
+						NPCLoader.OnHitPlayer(NPC, self, hit);
+
+						var hit2 = new Player.HurtModifiers();
+						NPCLoader.ModifyHitPlayer(NPC, self, ref hit2);
+
+						return;
+					}
+				}
+			}
+
+			orig(self);
+		}
+
+		public override bool PreDraw(SpriteBatch spriteBatch, Vector2 screenPos, Color drawColor)
+		{
+			return false;
 		}
 
 		public override void PostDraw(SpriteBatch spriteBatch, Vector2 screenPos, Color drawColor)
@@ -92,16 +135,6 @@ namespace StarlightRiver.Content.Bosses.SquidBoss
 				spriteBatch.Draw(glow, pos4 - screenPos, top.Frame(), color, 4.73f, top.Size() / 2, 1, 0, 0);
 				Lighting.AddLight(pos4, color.ToVector3() * 0.4f);
 			}
-		}
-
-		public override void ModifyHitPlayer(Player target, ref Player.HurtModifiers modifiers)
-		{
-			target.immune = true;
-			target.immuneTime = 1;
-
-			target.position.Y = NPC.position.Y + NPC.height;
-			target.velocity.Y += 5;
-			target.noKnockback = true;
 		}
 	}
 }
