+++ Content/Items/BaseTypes/CursedAccessory.cs
@@ -7,6 +7,7 @@ using Terraria.Audio;
 using Terraria.GameContent;
 using Terraria.Graphics.Effects;
 using Terraria.ID;
+using Terraria.UI;
 using Terraria.Utilities;
 
 namespace StarlightRiver.Content.Items.BaseTypes
@@ -65,7 +66,7 @@ namespace StarlightRiver.Content.Items.BaseTypes
 			Vector2 pos = position - origin * scale + new Vector2(0, 10) + Main.rand.NextVector2Circular(16, 16) + frame.Size() / 2f * scale;
 
 			if (!GoingBoom)
-				CursedAccessoryParticleManager.CursedSystem.AddParticle(pos, new Vector2(0, -0.4f), 0, 0.5f, CurseColor, 60, Vector2.Zero, alpha: 0);
+				CursedAccessorySystem.CursedSystem.AddParticle(pos, new Vector2(0, -0.4f), 0, 0.5f, CurseColor, 60, Vector2.Zero, alpha: 0);
 
 			drawpos = position - origin * scale + frame.Size() * scale * 0.5f;
 
@@ -73,7 +74,7 @@ namespace StarlightRiver.Content.Items.BaseTypes
 			{
 				float rot = Main.rand.NextFloat(6.28f);
 				float decay = 1f - boomTimer / 60f;
-				CursedAccessoryParticleManager.CursedSystem.AddParticle(drawpos + Vector2.One.RotatedBy(rot) * decay * 30, -Vector2.One.RotatedBy(rot) * 1.5f * decay, 0, 0.6f * boomTimer / 100f, new Color(255, Main.rand.Next(120, 200), 120) * 0.6f, 20, Vector2.Zero, alpha: 0);
+				CursedAccessorySystem.CursedSystem.AddParticle(drawpos + Vector2.One.RotatedBy(rot) * decay * 30, -Vector2.One.RotatedBy(rot) * 1.5f * decay, 0, 0.6f * boomTimer / 100f, new Color(255, Main.rand.Next(120, 200), 120) * 0.6f, 20, Vector2.Zero, alpha: 0);
 			}
 		}
 
@@ -85,7 +86,7 @@ namespace StarlightRiver.Content.Items.BaseTypes
 				SoundEngine.PlaySound(SoundID.Item123);
 
 				for (int k = 0; k <= 50; k++)
-					CursedAccessoryParticleManager.CursedSystem.AddParticle(drawpos, Vector2.One.RotatedByRandom(6.28f) * Main.rand.NextFloat(0.75f), 0, 1, CurseColor, 60, Vector2.Zero, alpha: 0);
+					CursedAccessorySystem.CursedSystem.AddParticle(drawpos, Vector2.One.RotatedByRandom(6.28f) * Main.rand.NextFloat(0.75f), 0, 1, CurseColor, 60, Vector2.Zero, alpha: 0);
 			}
 		}
 
@@ -116,17 +117,17 @@ namespace StarlightRiver.Content.Items.BaseTypes
 				{
 					float distance = Main.rand.NextFloat(4.25f);
 
-					CursedAccessoryParticleManager.CursedSystem.AddParticle(drawpos, Vector2.One.RotatedByRandom(6.28f) * Main.rand.NextFloat(4.25f, 4.75f), 1, 1.2f, CurseColor * 0.6f, 60, Vector2.Zero);
-					CursedAccessoryParticleManager.CursedSystem.AddParticle(drawpos, Vector2.One.RotatedByRandom(6.28f) * distance, 1, 0.8f, Color.Lerp(new Color(200, 60, 250, 0), CurseColor * 0.8f, distance / 4.25f), 60, Vector2.Zero);
+					CursedAccessorySystem.CursedSystem.AddParticle(drawpos, Vector2.One.RotatedByRandom(6.28f) * Main.rand.NextFloat(4.25f, 4.75f), 1, 1.2f, CurseColor * 0.6f, 60, Vector2.Zero);
+					CursedAccessorySystem.CursedSystem.AddParticle(drawpos, Vector2.One.RotatedByRandom(6.28f) * distance, 1, 0.8f, Color.Lerp(new Color(200, 60, 250, 0), CurseColor * 0.8f, distance / 4.25f), 60, Vector2.Zero);
 				}
 
-				CursedAccessoryParticleManager.ShardsSystem.SetTexture(tex);
+				CursedAccessorySystem.ShardsSystem.SetTexture(tex);
 
 				for (int x = 0; x < 5; x++)
 				{
 					for (int y = 0; y < 5; y++)
 					{
-						CursedAccessoryParticleManager.ShardsSystem.AddParticle(
+						CursedAccessorySystem.ShardsSystem.AddParticle(
 								drawpos - tex.Size() / 2f + new Vector2(x * tex.Width / 5f, y * tex.Height / 5f),
 								Vector2.One.RotatedByRandom(6.28f) * Main.rand.NextFloat(1, 2.55f),
 								0,
@@ -197,7 +198,7 @@ namespace StarlightRiver.Content.Items.BaseTypes
 		public virtual void SafeModifyTooltips(List<TooltipLine> tooltips) { }
 	}
 
-	class CursedAccessoryParticleManager : ModSystem
+	class CursedAccessorySystem : ModSystem
 	{
 		public static ParticleSystem CursedSystem;
 		public static ParticleSystem ShardsSystem;
@@ -206,12 +207,55 @@ namespace StarlightRiver.Content.Items.BaseTypes
 		{
 			CursedSystem = new ParticleSystem("StarlightRiver/Assets/Masks/GlowAlpha", UpdateCursedBody, ParticleSystem.AnchorOptions.UI);
 			ShardsSystem = new ParticleSystem("StarlightRiver/Assets/GUI/charm", UpdateShardsBody, ParticleSystem.AnchorOptions.UI);
+
+			On_ItemSlot.Draw_SpriteBatch_ItemArray_int_int_Vector2_Color += DrawSpecial;
 		}
 
 		public override void PostUpdateEverything()
 		{
 			CursedSystem?.UpdateParticles();
 			ShardsSystem?.UpdateParticles();
+		}
+
+		private void DrawSpecial(On_ItemSlot.orig_Draw_SpriteBatch_ItemArray_int_int_Vector2_Color orig, SpriteBatch sb, Item[] inv, int context, int slot, Vector2 position, Color color)
+		{
+			if (inv[slot].ModItem is CursedAccessory && context == 10)
+			{
+				Texture2D back = Assets.GUI.CursedBack.Value;
+				Color backcolor = (!Main.expertMode && slot == 8) ? Color.White * 0.25f : Color.White * 0.75f;
+
+				sb.Draw(back, position, null, backcolor, 0f, default, Main.inventoryScale, SpriteEffects.None, 0f);
+				RedrawItem(sb, inv, position, slot, color);
+			}
+			else
+			{
+				orig(sb, inv, context, slot, position, color);
+			}
+		}
+
+		//this is vanilla code. Only reasonable alternative is likely porting all drawing to IL.
+		internal static void RedrawItem(SpriteBatch sb, Item[] inv, Vector2 position, int slot, Color color)
+		{
+			Item item = inv[slot];
+			Vector2 scaleVector = Vector2.One * 52 * Main.inventoryScale;
+			Texture2D popupTex = Terraria.GameContent.TextureAssets.Item[item.type].Value;
+			Rectangle source = popupTex.Frame(1, 1, 0, 0);
+			Color currentColor = color;
+			float scaleFactor2 = 1f;
+			ItemSlot.GetItemLight(ref currentColor, ref scaleFactor2, item, false);
+			float scaleFactor = 1f;
+
+			if (source.Width > 32 || source.Height > 32)
+				scaleFactor = (source.Width <= source.Height) ? (32f / source.Height) : (32f / source.Width);
+
+			scaleFactor *= Main.inventoryScale;
+			Vector2 drawPos = position + scaleVector / 2f - source.Size() * scaleFactor / 2f;
+			Vector2 origin = source.Size() * (scaleFactor2 / 2f - 0.5f);
+
+			if (ItemLoader.PreDrawInInventory(item, sb, drawPos, source, item.GetAlpha(currentColor), item.GetColor(color), origin, scaleFactor * scaleFactor2))
+				sb.Draw(popupTex, drawPos, source, Color.White, 0f, origin, scaleFactor * scaleFactor2, SpriteEffects.None, 0f);
+
+			ItemLoader.PostDrawInInventory(item, sb, drawPos, source, item.GetAlpha(currentColor), item.GetColor(color), origin, scaleFactor * scaleFactor2);
 		}
 
 		private static void UpdateShardsBody(Particle particle)
