+++ Content/Tiles/Underground/EvasionShrine.cs
@@ -1,9 +1,9 @@
 ﻿using Microsoft.Xna.Framework.Graphics;
 using StarlightRiver.Content.Abilities;
-using StarlightRiver.Content.CustomHooks;
 using StarlightRiver.Content.Items.Misc;
 using StarlightRiver.Content.Tiles.Underground.EvasionShrineBullets;
 using StarlightRiver.Core.Systems.DummyTileSystem;
+using StarlightRiver.Core.Systems.NoBuildingSystem;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -99,7 +99,7 @@ namespace StarlightRiver.Content.Tiles.Underground
 
 			if (state != SHRINE_STATE_IDLE)
 			{
-				ProtectionWorld.AddRegionBySource(new Point16(ParentX, ParentY), ArenaTile);//stop calling this and call RemoveRegionBySource() when shrine is completed
+				NoBuildSystem.AddRegionBySource(new Point16(ParentX, ParentY), ArenaTile);//stop calling this and call RemoveRegionBySource() when shrine is completed
 
 				Dust.NewDustPerfect(Center + new Vector2(Main.rand.NextFloat(-24, 24), 28), ModContent.DustType<Dusts.Glow>(), Vector2.UnitY * -Main.rand.NextFloat(2), 0, new Color(150, 30, 205) * Windup, 0.2f);
 
@@ -132,7 +132,7 @@ namespace StarlightRiver.Content.Tiles.Underground
 							SpawnReward();
 							state = SHRINE_STATE_DEFEATED;
 							SetFrame(2);
-							ProtectionWorld.RemoveRegionBySource(new Point16(ParentX, ParentY));
+							NoBuildSystem.RemoveRegionBySource(new Point16(ParentX, ParentY));
 						}
 
 						return;
@@ -159,7 +159,7 @@ namespace StarlightRiver.Content.Tiles.Underground
 				{
 					state = SHRINE_STATE_IDLE;
 					attackOrder = null;
-					ProtectionWorld.RemoveRegionBySource(new Point16(ParentX, ParentY));
+					NoBuildSystem.RemoveRegionBySource(new Point16(ParentX, ParentY));
 				}
 
 				return;
