+++ Content/WorldGeneration/GenerateVitric.cs
@@ -1,6 +1,6 @@
-﻿using StarlightRiver.Content.CustomHooks;
-using StarlightRiver.Content.Tiles.Vitric;
+﻿using StarlightRiver.Content.Tiles.Vitric;
 using StarlightRiver.Content.Tiles.Vitric.Temple.GearPuzzle;
+using StarlightRiver.Core.Systems.NoBuildingSystem;
 using StarlightRiver.Noise;
 using System;
 using System.Collections.Generic;
@@ -46,7 +46,7 @@ namespace StarlightRiver.Core
 			//Basic biome information
 			vitricBiome = new Rectangle(GenVars.UndergroundDesertLocation.X - 25, GenVars.UndergroundDesertLocation.Y + GenVars.UndergroundDesertLocation.Height / 2, GenVars.UndergroundDesertLocation.Width + 50, vitricHeight);
 			//Boss arena protection
-			ProtectionWorld.ProtectedRegions.Add(VitricBossArena);
+			NoBuildSystem.protectedRegions.Add(VitricBossArena);
 
 			int minCeilingDepth = (int)(vitricBiome.Y + vitricBiome.Height / 2 - 17f * Math.Log(SLOPE_OFFSET - 8)); //Various informational variables - not to be changed
 			int maxCeilingDepth = minCeilingDepth + 7;
@@ -563,7 +563,7 @@ namespace StarlightRiver.Core
 
 			Point16 dims = StructureHelper.API.MultiStructureGenerator.GetStructureDimensions("Structures/VitricForge", StarlightRiver.Instance, 0);
 
-			ProtectionWorld.ProtectedRegions.Add(new Rectangle(x, vitricBiome.Center.Y - 10, dims.X, dims.Y));
+			NoBuildSystem.protectedRegions.Add(new Rectangle(x, vitricBiome.Center.Y - 10, dims.X, dims.Y));
 
 			NPC.NewNPC(new EntitySource_WorldGen(), (x + 80) * 16, (StarlightWorld.vitricBiome.Center.Y + 20) * 16, NPCType<Content.Bosses.GlassMiniboss.GlassweaverFriendly>());
 		}
